Moved and renamed logger
